discard deleed file post-merge
